Made some notes on node selection logic
Despite marking the node properly as visited, after the first iteration
the loop gets stuck because curNode never has a chance to properly
update itself after the first run. It only ever changes in the initial
loop selection.

Initial thoughts are to break the check in the if to first check if the
node has been visited before and then do the rest of the checks. We also
have to force curNode off of the first value temporarily after we update
the prev and dist arrays - when we loop back around the parent loop, the
node selection for loop can update the shortest node properly