Reworked graph structure
Old graph storage system was causing a lot of issues in terms of keeping
track which nodes were already visited. By including an 'id' member
we can better make use of the vector and remove elements as we visit
them while still keeping track of which node we're at. Before, IDs
were inferred from their position in the vector.